services/backend/requirements.txt
fastapi[all]
uvicorn
pydantic
passlib[bcrypt]
python-multipart
python-jose[cryptography]
psycopg[binary,pool]
